Extend identificaiton logic into 3 bytes and send capacity to client
The Intel datasheet is missleading when it says the identify
command replies with two bytes of padding and a one byte of
device ID: other datasheets state that the 3 bytes are manuf.
code, device ID and device capacity. So extend the identification
logic to use all 3 bytes.

Also made the capacity be sent back to the client, so it can size
a read operation to that size.

More cleanups needed: slow 1 byte transfers in the C code, and the
write client code could ensure the input isn't too big etc.