Added support for the Micron N25Q032A and rework polling for busy
This was figuired out by decoding with an LA what my minipro programmer
does, since all attempts at following the datasheet failed.

The poll for not busy was changed so instead of repeated reading of the
status register after requesting it once, it is requested each time
through the loop, with a 100ms wait (for erase only).

Protection bits are cleared by writing to the status register. This made
the KH25L3233F I filched from a picture frame work.

TODO: Speed up reads (they are slower then writes now because of reading
256 bytes at a time.