Extended programming to cater for programming iCE40UP FPGAs
This follows section 13 of the programming and configuration guide:

https://www.latticesemi.com/view_document?document_id=46502

That is, the FPGA is programmed directly via the SPI slave mode.

Further speed optimisation is no doubt possible, etc. Also uses
the /CRESET pin in the normal flash case, so the FPGA can be reset,
after programming, if one is present. This is currently done on
reads of the flash as well, which is probably not right.